+++ src/App.js
@@ -2,15 +2,15 @@ import React, { useState, useEffect } from 'react';
 import { BrowserRouter, Route, Routes } from 'react-router-dom';
 import './App.css';
 
-import About from './components/About';
-import Activities from './components/Activities';
-import Contact from './components/Contact';
-import Contribute from './components/Contribute';
-import Error404 from './components/Error404';
-import Home from './components/Home';
+import About from './pages/About';
+import Activities from './pages/Activities';
+import Contact from './pages/Contact';
+import Contribute from './pages/Contribute';
+import Error404 from './pages/Error404';
+import Home from './pages/Home';
 import Loading from './components/Loading';
-import News from './components/News';
-import Testimonials from './components/Testimonials';
+import News from './pages/News';
+import Testimonials from './pages/Testimonials';
 
 function App() {
   const [isLoading, setIsLoading] = useState(true);
@@ -18,26 +18,25 @@ function App() {
   useEffect(() => {
     setTimeout(() => {
       setIsLoading(false);
-    }, 1200)
-  })
+    }, 1200);
+  });
   return (
     <BrowserRouter>
-    {
-      isLoading===true?
-      <Loading />
-      :
-      <Routes>
-        <Route path='/' element={<Home/>} />
-        <Route path='/about' element={<About/>} />
-        <Route path='/activities' element={<Activities/>} />
-        <Route path='/news' element={<News/>} />
-        <Route path='/testimonials' element={<Testimonials/>} />
-        <Route path='/contact' element={<Contact/>} />
-        <Route path='/contribute' element={<Contribute/>} />
-        {/* returns the Erorr404 component in case the path does't exist */}
-        <Route path='*' element={<Error404/>} />
-      </Routes>
-    }
+      {isLoading === true ? (
+        <Loading />
+      ) : (
+        <Routes>
+          <Route path="/" element={<Home />} />
+          <Route path="/about" element={<About />} />
+          <Route path="/activities" element={<Activities />} />
+          <Route path="/news" element={<News />} />
+          <Route path="/testimonials" element={<Testimonials />} />
+          <Route path="/contact" element={<Contact />} />
+          <Route path="/contribute" element={<Contribute />} />
+          {/* returns the Erorr404 component in case the path does't exist */}
+          <Route path="*" element={<Error404 />} />a
+        </Routes>
+      )}
     </BrowserRouter>
   );
 }
