+++ src/pages/Activities.js
