+++ src/pages/News.js
@@ -1,0 +1,11 @@
+import React from 'react'
+
+function News() {
+    return (
+        <>
+            <h1>News...</h1>    
+        </>
+    )
+}
+
+export default News
